requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
jinja2
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
gunicorn
